+++ internal/sdk/models/operations/listbouncers.go
@@ -17,9 +17,9 @@ type ListBouncersRequest struct {
 	// The name of the branch that owns this resource
 	Branch string `pathParam:"style=simple,explode=false,name=branch"`
 	// If provided, specifies the page offset of returned results
-	Page *float64 `default:"1" queryParam:"style=form,explode=true,name=page"`
+	Page *int64 `default:"1" queryParam:"style=form,explode=true,name=page"`
 	// If provided, specifies the number of returned results
-	PerPage *float64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
+	PerPage *int64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
 }
 
 func (l ListBouncersRequest) MarshalJSON() ([]byte, error) {
@@ -54,14 +54,14 @@ func (l *ListBouncersRequest) GetBranch() string {
 	return l.Branch
 }
 
-func (l *ListBouncersRequest) GetPage() *float64 {
+func (l *ListBouncersRequest) GetPage() *int64 {
 	if l == nil {
 		return nil
 	}
 	return l.Page
 }
 
-func (l *ListBouncersRequest) GetPerPage() *float64 {
+func (l *ListBouncersRequest) GetPerPage() *int64 {
 	if l == nil {
 		return nil
 	}
