+++ internal/sdk/models/operations/listdatabaseregions.go
@@ -13,9 +13,9 @@ type ListDatabaseRegionsRequest struct {
 	// The name of the database
 	Database string `pathParam:"style=simple,explode=false,name=database"`
 	// If provided, specifies the page offset of returned results
-	Page *float64 `default:"1" queryParam:"style=form,explode=true,name=page"`
+	Page *int64 `default:"1" queryParam:"style=form,explode=true,name=page"`
 	// If provided, specifies the number of returned results
-	PerPage *float64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
+	PerPage *int64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
 }
 
 func (l ListDatabaseRegionsRequest) MarshalJSON() ([]byte, error) {
@@ -43,14 +43,14 @@ func (l *ListDatabaseRegionsRequest) GetDatabase() string {
 	return l.Database
 }
 
-func (l *ListDatabaseRegionsRequest) GetPage() *float64 {
+func (l *ListDatabaseRegionsRequest) GetPage() *int64 {
 	if l == nil {
 		return nil
 	}
 	return l.Page
 }
 
-func (l *ListDatabaseRegionsRequest) GetPerPage() *float64 {
+func (l *ListDatabaseRegionsRequest) GetPerPage() *int64 {
 	if l == nil {
 		return nil
 	}
